@@ -3,10 +3,11 @@
 buildscript {
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3'
-        classpath 'com.google.gms:google-services:3.2.0'
+        classpath 'com.android.tools.build:gradle:3.4.2'
+        classpath 'com.google.gms:google-services:4.3.0'
     }
 }
 
@@ -16,18 +17,22 @@ allprojects {
         maven {
             url 'https://maven.google.com'
         }
+        maven {
+            // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
+            url("$rootDir/../node_modules/react-native/android")
+        }
+        google()
     }
 }
 
 apply plugin: "com.android.library"
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.3"
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
         ndk {
@@ -45,10 +50,9 @@ android {
 
 dependencies {
     compile fileTree(dir: "libs", include: ["*.jar"])
-    compile "com.android.support:appcompat-v7:23.0.1"
-    compile "com.facebook.react:react-native:+"  // From node_modules
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation "com.facebook.react:react-native:+"  // From node_modules
 
-    compile 'com.google.firebase:firebase-messaging:12.0.1'
-    compile 'com.google.firebase:firebase-core:12.0.1'
-    compile 'com.firebase:firebase-jobdispatcher:0.6.0'
+    implementation 'com.google.firebase:firebase-messaging:+'
+    implementation 'com.google.firebase:firebase-core:+'
 }
