@@ -14,24 +14,18 @@
 package com.amazonaws.amplify.pushnotification;
 
 import android.util.Log;
-import android.os.Bundle;
 import android.app.Application;
 import android.content.IntentFilter;
-import android.content.BroadcastReceiver;
 
-import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.Callback;
 
-import com.facebook.react.ReactApplication;
-import com.facebook.react.ReactInstanceManager;
-
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.iid.InstanceIdResult;
 
-import com.amazonaws.amplify.pushnotification.modules.RNPushNotificationJsDelivery;
 import com.amazonaws.amplify.pushnotification.modules.RNPushNotificationBroadcastReceiver;
 
 public class RNPushNotificationModule extends ReactContextBaseJavaModule {
@@ -64,9 +58,14 @@ public class RNPushNotificationModule extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
-    public void getToken(Callback callback) {
-        String token =  FirebaseInstanceId.getInstance().getToken();
-        Log.i(LOG_TAG, "getting token" + token);
-        callback.invoke(token);
+    public void getToken(final Callback callback) {
+        FirebaseInstanceId.getInstance().getInstanceId().addOnSuccessListener(new OnSuccessListener<InstanceIdResult>() {
+            @Override
+            public void onSuccess(InstanceIdResult instanceIdResult) {
+                String newToken = instanceIdResult.getToken();
+                Log.e("newToken", newToken);
+                callback.invoke(newToken);
+            }
+        });
     }
 }
